requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-Migrate==4.0.4
psycopg2-binary==2.9.6
python-dotenv==1.0.0
Werkzeug==2.3.6
email_validator==2.0.0.post2
numpy>=1.24.0
reportlab>=4.0.0
matplotlib>=3.7.0
WTForms>=3.0.0
Flask-WTF>=1.1.0
Pillow>=9.0.0
gunicorn>=21.0.0
